Fixes for pytest discovery output plugin
